Register function and other stuff
Implemented a function for registering as a new customer.
had to do small changes in the start-menu in program.cs and in the file
creation-function in files.cs. The loop didn't exit as it should, i
changed it to a while(true) to work better. The file created an admin
but forgot to add a new line so the first customer to be created ended
up on the same line. Fixed it by adding environment.newline.

also added the function for masking the password being put in as it made
sense at this stage.

and created a Utils-class for functions we might want to use often bt
doesnt fit in one class. for now there's only a funcion that asks the
user for a specific input and then saves that to a string.

This commit got a lot bigger than intended, sorry.