Login function
Added a login-function that should be working once everything has been
merged.

Had to ad a constructor in the product class to be able to load a cart
for every customer that logs in. Also removed the Cart-class and just
made it so that a cart is a List<Product> in the customer-constructor.
Followed Manuels example code that he put up on the blogg.