Function for viewing order history
Added function for viewing order history.

for this to work we need to save orders in a specific manner. Like this:
row 1: productName,productprice
row 2: productName,productprice
.
.
.
Row x: Sum,<total sum of order in numbers>
row y: Timestamp,YYYY/MM//DD HH:MM

I created an enum with the reciept tags: Sum and Timestamp to use when
creating a reciept for row X and Y.